Behaviour Graphs: replace the C++ script-behaviour system with visual scripting
Full replacement of Zenith_ScriptComponent / Zenith_ScriptAsset /
Zenith_ScriptBehaviour / .zscript with a designer-facing Behaviour Graph
system. No backwards compatibility; the old system is deleted outright
(zero-remnant grep gate: 0 hits incl. registered-name strings).

New engine surface:
- Zenith/Core property system: ZENITH_PROPERTY macros, tagged
  Zenith_PropertyValue, fn-ptr get/set thunks, name+type-matched framed
  serialization, auto ImGui panels, .ztune live-tuning reload.
- Zenith/Scripting: .bgraph graph runtime (event-driven exec-chain
  interpreter, SUCCESS/FAILURE/RUNNING, blackboards, unresolved-node
  preservation, per-node type versions). ZenithBase+ZenithECS deps only.
- Zenith_BehaviourGraphAsset (.bgraph, versioned), Zenith_GraphComponent
  (slot list + blackboard overrides, meta order 60), ~24 engine nodes,
  FileWatcher hot reload drained at the main-loop safe point.
- Hand-rolled node editor panel (palette/canvas/pins/properties/variables,
  live execution highlight, unresolved nodes error-red) + ImGui simulated-
  input bridge (inject between glfw NewFrame and ImGui::NewFrame) +
  ZENITH_TESTING screen-pos accessors for simulator-driven UI tests.
- Component-meta registry gains concept-detected OnCollision* hooks;
  physics dispatches collisions through the registry; post-seal
  registration re-finalizes (game components register from project hooks).
- Android: .bgraph rides the existing APK asset pipeline; agde Debug +
  Release link.

All 11 games migrated to game ECS components (ZENITH_REGISTER_COMPONENT)
+ graphs; old *_Behaviour files and .zscript assets deleted; the
Project_RegisterScriptBehaviours hook renamed Project_RegisterGameComponents.
The orphaned Test game was restored to the build (the Sharpmake class
formerly named TestGameProject actually generated Sokoban; renamed to
SokobanGameProject and a real TestGameProject added).

Notable fixes en route:
- Priest froze mid-pursuit after migration: old behaviours dispatched at
  order 60 (before AIAgent=90), components run at 100+ - the per-frame
  body-velocity zero now runs AFTER the nav tick and erased it. Zero only
  when the nav agent has no path.
- Scenes created SCENE_LOAD_ADDITIVE_WITHOUT_LOADING never update, so
  components added there never receive the awake wave-drain; affected
  test fixtures call OnAwake() explicitly (exactly-once by construction).
- A unit test resolved the meta registry by the deleted "Script" display
  name - null-deref that masqueraded as heap corruption in an unrelated
  test; found via an ASan build.

Verification (all on this tree): unit suites 716/716 (DP exe) and 721/721
(RenderTest exe); DP headless suite 138/138 incl. the new windowed
Test_GraphEditorScreenshotTour; CB suite 45/45 + CB_HumanSession
(2930 frames); windowed Test_GraphEditorLiveAuthoring flagship (mouse-
authored graph -> save -> play -> live mid-play edit + hot reload),
Night1WinGolden, PersonalityPlaythrough_Casual (9416 frames); Vulkan
Debug/Release x True/False + D3D12 null backend + all 11 games build;
complexity + architecture/lints gates green; 9-agent adversarial review
of the final diff confirmed no functional defects.

+++ Games/DevilsPlayground/Tests/Test_P1Save_RoundTripMeta.cpp
@@ -100,7 +100,7 @@ static void Setup_P1SaveRoundTrip()
 	const bool bOk = DP_Save::TryLoad(xStream, xLoaded);
 
 	std::printf("[P1SaveRoundTrip] wrote=%llu bytes, TryLoad=%d\n",
-		ulWritten, (int)bOk);
+		static_cast<unsigned long long>(ulWritten), (int)bOk);
 
 	if (!bOk)
 	{
